Build cleanly with warnings as errors
A clean build with -Werror turned up four things the ordinary build had let
through: two narrowing conversions from Qt's 64-bit size type, several test
calls whose return value was intentionally discarded without saying so, and
a lambda parameter shadowing the one it was nested in.

None of them was a live defect, but the debug preset and CI both build with
-Werror, so leaving them would have meant a build that fails for the next
person on their first try.

+++ src/core/rewrite/formats/SqliteRewriter.cpp
@@ -123,7 +123,8 @@ QList<RewriteEdit> rewriteSqlite(const QString& path, const QString& table, cons
     for (const Update& row : updates) {
         const QByteArray value = row.newValue.toUtf8();
         sqlite3_reset(update.get());
-        sqlite3_bind_text(update.get(), 1, value.constData(), value.size(), SQLITE_TRANSIENT);
+        sqlite3_bind_text(update.get(), 1, value.constData(), static_cast<int>(value.size()),
+                          SQLITE_TRANSIENT);
         sqlite3_bind_int64(update.get(), 2, row.rowId);
 
         if (sqlite3_step(update.get()) != SQLITE_DONE) {
